+++ Sources/OmFileFormatC/src/om_encoder.c
@@ -15,7 +15,7 @@
 
 
 // Initialize om_file_encoder
-void om_encoder_init(om_encoder_t* encoder, float scale_factor, float add_offset, om_compression_t compression, om_datatype_t data_type, const uint64_t* dimensions, const uint64_t* chunks, uint64_t dimension_count, uint64_t lut_chunk_element_count) {
+void om_encoder_init(om_encoder_t* encoder, float scale_factor, float add_offset, om_compression_t compression, om_datatype_t data_type, const size_t* dimensions, const size_t* chunks, size_t dimension_count, size_t lut_chunk_element_count) {
     encoder->scale_factor = scale_factor;
     encoder->add_offset = add_offset;
     encoder->dimensions = dimensions;
@@ -63,16 +63,16 @@ void om_encoder_init(om_encoder_t* encoder, float scale_factor, float add_offset
 }
 
 // Calculate number of chunks
-uint64_t om_encoder_number_of_chunks(const om_encoder_t* encoder) {
-    uint64_t n = 1;
+size_t om_encoder_number_of_chunks(const om_encoder_t* encoder) {
+    size_t n = 1;
     for (int i = 0; i < encoder->dimension_count; i++) {
         n *= divide_rounded_up(encoder->dimensions[i], encoder->chunks[i]);
     }
     return n;
 }
 
-uint64_t om_encoder_number_of_chunks_in_array(const om_encoder_t* encoder, const uint64_t* array_count) {
-    uint64_t numberOfChunksInArray = 1;
+size_t om_encoder_number_of_chunks_in_array(const om_encoder_t* encoder, const size_t* array_count) {
+    size_t numberOfChunksInArray = 1;
     for (int i = 0; i < encoder->dimension_count; i++) {
         numberOfChunksInArray *= divide_rounded_up(array_count[i], encoder->chunks[i]);
     }
@@ -80,8 +80,8 @@ uint64_t om_encoder_number_of_chunks_in_array(const om_encoder_t* encoder, const
 }
 
 // Calculate chunk buffer size
-uint64_t om_encoder_compress_chunk_buffer_size(const om_encoder_t* encoder) {
-    uint64_t chunkLength = 1;
+size_t om_encoder_compress_chunk_buffer_size(const om_encoder_t* encoder) {
+    size_t chunkLength = 1;
     for (int i = 0; i < encoder->dimension_count; i++) {
         chunkLength *= encoder->chunks[i];
     }
@@ -90,55 +90,55 @@ uint64_t om_encoder_compress_chunk_buffer_size(const om_encoder_t* encoder) {
 }
 
 // Calculate the size of the compressed LUT buffer
-uint64_t om_encoder_compress_lut_buffer_size(const om_encoder_t* encoder, const uint64_t* lookUpTable, uint64_t lookUpTableCount) {
+size_t om_encoder_compress_lut_buffer_size(const om_encoder_t* encoder, const size_t* lookUpTable, size_t lookUpTableCount) {
     unsigned char buffer[MAX_LUT_ELEMENTS+32] = {0};
-    uint64_t nLutChunks = divide_rounded_up(lookUpTableCount, encoder->lut_chunk_element_count);
-    uint64_t maxLength = 0;
+    size_t nLutChunks = divide_rounded_up(lookUpTableCount, encoder->lut_chunk_element_count);
+    size_t maxLength = 0;
     for (int i = 0; i < nLutChunks; i++) {
-        uint64_t rangeStart = i * encoder->lut_chunk_element_count;
-        uint64_t rangeEnd = min(rangeStart + encoder->lut_chunk_element_count, lookUpTableCount);
+        size_t rangeStart = i * encoder->lut_chunk_element_count;
+        size_t rangeEnd = min(rangeStart + encoder->lut_chunk_element_count, lookUpTableCount);
         size_t len = p4ndenc64((uint64_t*)&lookUpTable[rangeStart], rangeEnd - rangeStart, buffer);
         if (len > maxLength) maxLength = len;
     }
     /// Compression function can write 32 integers more
-    return maxLength * nLutChunks + 32 * sizeof(uint64_t);
+    return maxLength * nLutChunks + 32 * sizeof(size_t);
 }
 
 // Compress the LUT to an output buffer and return the size of the compressed LUT
-uint64_t om_encoder_compress_lut(const om_encoder_t* encoder, const uint64_t* lookUpTable, uint64_t lookUpTableCount, uint8_t* out, uint64_t compressed_lut_buffer_size) {
-    uint64_t nLutChunks = divide_rounded_up(lookUpTableCount, encoder->lut_chunk_element_count);
-    uint64_t lutSize = compressed_lut_buffer_size - 32 * sizeof(uint64_t);
-    uint64_t lutChunkLength = lutSize / nLutChunks;
+size_t om_encoder_compress_lut(const om_encoder_t* encoder, const size_t* lookUpTable, size_t lookUpTableCount, uint8_t* out, size_t compressed_lut_buffer_size) {
+    size_t nLutChunks = divide_rounded_up(lookUpTableCount, encoder->lut_chunk_element_count);
+    size_t lutSize = compressed_lut_buffer_size - 32 * sizeof(size_t);
+    size_t lutChunkLength = lutSize / nLutChunks;
 
-    for (uint64_t i = 0; i < nLutChunks; i++) {
-        uint64_t rangeStart = i * encoder->lut_chunk_element_count;
-        uint64_t rangeEnd = min(rangeStart + encoder->lut_chunk_element_count, lookUpTableCount);
+    for (size_t i = 0; i < nLutChunks; i++) {
+        size_t rangeStart = i * encoder->lut_chunk_element_count;
+        size_t rangeEnd = min(rangeStart + encoder->lut_chunk_element_count, lookUpTableCount);
         p4ndenc64((uint64_t*)&lookUpTable[rangeStart], rangeEnd - rangeStart, &out[i * lutChunkLength]);
     }
     return lutSize;
 }
 
-size_t om_encoder_compress_chunk(const om_encoder_t* encoder, const void* array, const uint64_t* arrayDimensions, const uint64_t* arrayOffset, const uint64_t* arrayCount, uint64_t chunkIndex, uint64_t chunkIndexOffsetInThisArray, uint8_t* out, uint8_t* chunkBuffer) {
+size_t om_encoder_compress_chunk(const om_encoder_t* encoder, const void* array, const size_t* arrayDimensions, const size_t* arrayOffset, const size_t* arrayCount, size_t chunkIndex, size_t chunkIndexOffsetInThisArray, uint8_t* out, uint8_t* chunkBuffer) {
     /// The total size of `arrayDimensions`. Only used to check for out of bound reads
-    uint64_t arrayTotalCount = 1;
-    for (uint64_t i = 0; i < encoder->dimension_count; i++) {
+    size_t arrayTotalCount = 1;
+    for (size_t i = 0; i < encoder->dimension_count; i++) {
         arrayTotalCount *= arrayDimensions[i];
     }
     
-    uint64_t rollingMultiply = 1;
-    uint64_t rollingMultiplyChunkLength = 1;
-    uint64_t rollingMultiplyTargetCube = 1;
-    uint64_t readCoordinate = 0;
-    uint64_t writeCoordinate = 0;
-    uint64_t linearReadCount = 1;
+    size_t rollingMultiply = 1;
+    size_t rollingMultiplyChunkLength = 1;
+    size_t rollingMultiplyTargetCube = 1;
+    size_t readCoordinate = 0;
+    size_t writeCoordinate = 0;
+    size_t linearReadCount = 1;
     bool linearRead = true;
-    uint64_t lengthLast = 0;
+    size_t lengthLast = 0;
 
     for (int64_t i = encoder->dimension_count - 1; i >= 0; i--) {
-        uint64_t nChunksInThisDimension = divide_rounded_up(encoder->dimensions[i], encoder->chunks[i]);
-        uint64_t c0 = (chunkIndex / rollingMultiply) % nChunksInThisDimension;
-        uint64_t c0Offset = (chunkIndexOffsetInThisArray / rollingMultiply) % nChunksInThisDimension;
-        uint64_t length0 = min((c0 + 1) * encoder->chunks[i], encoder->dimensions[i]) - c0 * encoder->chunks[i];
+        size_t nChunksInThisDimension = divide_rounded_up(encoder->dimensions[i], encoder->chunks[i]);
+        size_t c0 = (chunkIndex / rollingMultiply) % nChunksInThisDimension;
+        size_t c0Offset = (chunkIndexOffsetInThisArray / rollingMultiply) % nChunksInThisDimension;
+        size_t length0 = min((c0 + 1) * encoder->chunks[i], encoder->dimensions[i]) - c0 * encoder->chunks[i];
 
         if (i == encoder->dimension_count - 1) {
             lengthLast = length0;
@@ -163,7 +163,7 @@ size_t om_encoder_compress_chunk(const om_encoder_t* encoder, const void* array,
         rollingMultiplyChunkLength *= length0;
     }
 
-    uint64_t lengthInChunk = rollingMultiplyChunkLength;
+    size_t lengthInChunk = rollingMultiplyChunkLength;
 
     while (true) {
         assert(readCoordinate + linearReadCount <= arrayTotalCount);
@@ -179,8 +179,8 @@ size_t om_encoder_compress_chunk(const om_encoder_t* encoder, const void* array,
         linearReadCount = 1;
 
         for (int64_t i = encoder->dimension_count - 1; i >= 0; i--) {
-            uint64_t qPos = ((readCoordinate / rollingMultiplyTargetCube) % arrayDimensions[i] - arrayOffset[i]) / encoder->chunks[i];
-            uint64_t length0 = min((qPos + 1) * encoder->chunks[i], arrayCount[i]) - qPos * encoder->chunks[i];
+            size_t qPos = ((readCoordinate / rollingMultiplyTargetCube) % arrayDimensions[i] - arrayOffset[i]) / encoder->chunks[i];
+            size_t length0 = min((qPos + 1) * encoder->chunks[i], arrayCount[i]) - qPos * encoder->chunks[i];
             readCoordinate += rollingMultiplyTargetCube;
 
             if (i == encoder->dimension_count - 1 && !(arrayCount[i] == length0 && arrayDimensions[i] == length0)) {
@@ -192,7 +192,7 @@ size_t om_encoder_compress_chunk(const om_encoder_t* encoder, const void* array,
             } else {
                 linearRead = false;
             }
-            uint64_t q0 = ((readCoordinate / rollingMultiplyTargetCube) % arrayDimensions[i] - arrayOffset[i]) % encoder->chunks[i];
+            size_t q0 = ((readCoordinate / rollingMultiplyTargetCube) % arrayDimensions[i] - arrayOffset[i]) % encoder->chunks[i];
             if (q0 != 0 && q0 != length0) {
                 break;
             }
