+++ Tests/OmFileFormatTests/OmFileFormatTests.swift
@@ -192,7 +192,7 @@ final class OmFileFormatTests: XCTestCase {
         let file = "writetest.om"
         try FileManager.default.removeItemIfExists(at: file)
         
-        let dims = [UInt64(3),3,3]
+        let dims = [3,3,3]
         let writer = OmFileWriterArray(dimensions: dims, chunkDimensions: [2,2,2], compression: .p4nzdec256, datatype: .float, scale_factor: 1, add_offset: 0)
         let buffer = OmWriteBuffer(capacity: 1)
         let fn = try FileHandle.createNewFile(file: file)
@@ -239,7 +239,7 @@ final class OmFileFormatTests: XCTestCase {
         let file = "writetest.om"
         try FileManager.default.removeItemIfExists(at: file)
         
-        let dims = [UInt64(5),5]
+        let dims = [5,5]
         let writer = OmFileWriterArray(dimensions: dims, chunkDimensions: [2,2], compression: .p4nzdec256, datatype: .float, scale_factor: 1, add_offset: 0, lutChunkElementCount: 2)
         let buffer = OmWriteBuffer(capacity: 1)
         let fn = try FileHandle.createNewFile(file: file)
@@ -340,7 +340,7 @@ final class OmFileFormatTests: XCTestCase {
         let file = "writetest.om"
         try FileManager.default.removeItemIfExists(at: file)
         
-        let dims = [UInt64(5),5]
+        let dims = [5,5]
         let writer = OmFileWriterArray(dimensions: dims, chunkDimensions: [2,2], compression: .p4nzdec256, datatype: .float, scale_factor: 1, add_offset: 0, lutChunkElementCount: 2)
         let buffer = OmWriteBuffer(capacity: 1)
         let fn = try FileHandle.createNewFile(file: file)
